web/msfrognymize2/app/requirements.txt
opencv-python
flask
dlib
numpy
cryptography
pillow
piexif
gunicorn
redisbeat
celery
redis
requests
celery-redbeat